+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project are documented below.
 
 The format is based on [keep a changelog](http://keepachangelog.com) and this project uses [semantic versioning](http://semver.org).
 
+## [Unreleased]
+
+
 ## [3.7.0] - 2021-09-28
 ### Added
 - New config options to enforce a single socket per user, and a single match per socket.